Develop

+++ src/main/java/com/ssafy/algogo/submission/service/impl/SubmissionServiceImpl.java
@@ -5,18 +5,23 @@ import com.ssafy.algogo.common.advice.ErrorCode;
 import com.ssafy.algogo.common.utils.S3Service;
 import com.ssafy.algogo.problem.entity.ProgramProblem;
 import com.ssafy.algogo.problem.repository.ProgramProblemRepository;
+import com.ssafy.algogo.review.repository.RequireReviewRepository;
+import com.ssafy.algogo.submission.dto.ReviewRematchTargetQueryDto;
 import com.ssafy.algogo.submission.dto.request.SubmissionRequestDto;
 import com.ssafy.algogo.submission.dto.request.UserSubmissionRequestDto;
 import com.ssafy.algogo.submission.dto.response.SubmissionListResponseDto;
 import com.ssafy.algogo.submission.dto.response.SubmissionResponseDto;
+import com.ssafy.algogo.submission.dto.response.TrendIdsResponseDto;
 import com.ssafy.algogo.submission.dto.response.UserSubmissionPageResponseDto;
 import com.ssafy.algogo.submission.dto.response.UserSubmissionResponseDto;
 import com.ssafy.algogo.submission.entity.Algorithm;
 import com.ssafy.algogo.submission.entity.Submission;
 import com.ssafy.algogo.submission.entity.SubmissionAlgorithm;
+import com.ssafy.algogo.submission.event.SubmissionEvent;
 import com.ssafy.algogo.submission.repository.AlgorithmRepository;
 import com.ssafy.algogo.submission.repository.SubmissionAlgorithmRepository;
 import com.ssafy.algogo.submission.repository.SubmissionRepository;
+import com.ssafy.algogo.submission.service.ReviewMatchService;
 import com.ssafy.algogo.submission.service.SubmissionService;
 import com.ssafy.algogo.user.entity.User;
 import com.ssafy.algogo.user.repository.UserRepository;
@@ -24,6 +29,7 @@ import java.util.HashSet;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.context.ApplicationEventPublisher;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -40,18 +46,22 @@ public class SubmissionServiceImpl implements SubmissionService {
     private final ProgramProblemRepository programProblemRepository;
     private final SubmissionAlgorithmRepository submissionAlgorithmRepository;
     private final AlgorithmRepository algorithmRepository;
+    private final RequireReviewRepository requireReviewRepository;
 
+    private final ApplicationEventPublisher applicationEventPublisher;
+    private final ReviewMatchService reviewMatchService;
     private final S3Service s3Service;
 
     @Override
-    @Transactional(readOnly = true)
     public SubmissionResponseDto getSubmission(Long submissionId) {
-        Submission submission = submissionRepository.findById(submissionId)
-            .orElseThrow(
-                () -> new CustomException("제출 정보가 잘못 되었습니다.", ErrorCode.INVALID_PARAMETER));
+        Submission submission = submissionRepository.findById(submissionId).orElseThrow(
+            () -> new CustomException("존재하지 않는 제출입니다.", ErrorCode.SUBMISSION_NOT_FOUND));
 
         List<Algorithm> usedAlgorithmList = algorithmRepository.findAllAlgorithmsBySubmissionId(
             submission.getId());
+
+        submission.increaseViewCount();
+
         return SubmissionResponseDto.from(submission, usedAlgorithmList);
     }
 
@@ -63,15 +73,14 @@ public class SubmissionServiceImpl implements SubmissionService {
             .orElseThrow(() -> new CustomException("존재하지 않는 회원입니다.", ErrorCode.USER_NOT_FOUND));
         ProgramProblem programProblem = programProblemRepository.findById(
             submissionRequestDto.getProgramProblemId()).orElseThrow(
-            () -> new CustomException("프로그램 문제 정보가 잘못 되었습니다.", ErrorCode.INVALID_PARAMETER));
+            () -> new CustomException("존재하지 않는 프로그램 문제입니다.", ErrorCode.PROGRAM_PROBLEM_NOT_FOUND));
 
         // 코드는 S3에 저장
         String s3CodeUrl = s3Service.uploadText(userId, submissionRequestDto.getCode());
 
         // 제출 저장
         Submission submission = submissionRepository.save(
-            Submission.builder().language(submissionRequestDto.getLanguage())
-                .code(s3CodeUrl)
+            Submission.builder().language(submissionRequestDto.getLanguage()).code(s3CodeUrl)
                 .execTime(submissionRequestDto.getExecTime())
                 .memory(submissionRequestDto.getMemory())
                 .strategy(submissionRequestDto.getStrategy())
@@ -87,8 +96,15 @@ public class SubmissionServiceImpl implements SubmissionService {
         }
 
         // 제출 시 사용한 알고리즘 저장
-        List<Algorithm> usedAlgorithmList = createSubmissionAlgorithmAndFetch(
-            submissionRequestDto, submission);
+        List<Algorithm> usedAlgorithmList = createSubmissionAlgorithmAndFetch(submissionRequestDto,
+            submission);
+
+//        // 리뷰 매칭
+//        reviewMatchService.matchReviewers(submission, usedAlgorithmList, 2);
+
+        // 비동기 적용을 위해 Event 발행
+        applicationEventPublisher.publishEvent(
+            new SubmissionEvent(submission, usedAlgorithmList, 2));
 
         return SubmissionResponseDto.from(submission, usedAlgorithmList);
     }
@@ -96,34 +112,33 @@ public class SubmissionServiceImpl implements SubmissionService {
     @Override
     public void deleteSubmission(Long userId, Long submissionId) {
         User user = userRepository.findById(userId)
-            .orElseThrow(
-                () -> new CustomException("존재하지 않는 회원입니다.", ErrorCode.USER_NOT_FOUND));
+            .orElseThrow(() -> new CustomException("존재하지 않는 회원입니다.", ErrorCode.USER_NOT_FOUND));
 
-        Submission submission = submissionRepository.findById(submissionId)
-            .orElseThrow(
-                () -> new CustomException("잘못된 제출 ID 정보가 포함되어 있습니다.", ErrorCode.BAD_REQUEST));
+        Submission submission = submissionRepository.findById(submissionId).orElseThrow(
+            () -> new CustomException("존재하지 않는 제출입니다.", ErrorCode.SUBMISSION_NOT_FOUND));
 
-        // submission <-> review의 삭제는 어플리케이션 레벨에서 먼저 지운 후 지우는게 좋을 것 같음
-        // onDelete를 사용하기에는 DB레벨에 보내버리게 되기도하고
-        // 요구된 리뷰 부분을 삭제할 때 비즈니스 로직이 추가되니까
+        if (!submission.getUser().equals(user)) {
+            throw new CustomException("제출과 회원 정보가 일치하지 않습니다.", ErrorCode.INVALID_PARAMETER);
+        }
 
-        // ** 로직 설계
-        // 요구된 리뷰 조회
-        // 수행여부 x시 수정 필요
-        // 요구된 리뷰 삭제
-        // 리뷰 리매칭 요청
-
+        // 요구된 리뷰 중 수행 X 조회
+        List<ReviewRematchTargetQueryDto> reviewRematchTargetList = requireReviewRepository.findAllReviewRematchTargets(
+            submission.getId());
+        // 원 제출 삭제 -> 요구된 리뷰 삭제(@OnCascade.DELETE)
         submissionRepository.delete(submission);
+        // 리뷰 리매칭
+        for (ReviewRematchTargetQueryDto target : reviewRematchTargetList) {
+            reviewMatchService.matchReviewers(target.submission(), target.algorithmList(), 1);
+        }
     }
 
     private List<Algorithm> createSubmissionAlgorithmAndFetch(
-        SubmissionRequestDto submissionRequestDto,
-        Submission submission) {
+        SubmissionRequestDto submissionRequestDto, Submission submission) {
         List<Long> usedAlgorithmRequestIdList = submissionRequestDto.getAlgorithmList();
 
         // SubmissionAlgorithm create 요청 알고리즘 검증을 위한 조회
-        List<Algorithm> usedAlgorithmList = algorithmRepository.findAllById(new HashSet<>(
-            usedAlgorithmRequestIdList));
+        List<Algorithm> usedAlgorithmList = algorithmRepository.findAllById(
+            new HashSet<>(usedAlgorithmRequestIdList));
 
         // 중복 제거 후 사이즈와 실제 DB 조회 사이즈가 다르면 없는 알고리즘이 요청에 섞인 경우임.
         if (new HashSet<>(usedAlgorithmRequestIdList).size() != usedAlgorithmList.size()) {
@@ -142,43 +157,44 @@ public class SubmissionServiceImpl implements SubmissionService {
     @Transactional(readOnly = true)
     public SubmissionListResponseDto getSubmissionHistories(Long userId, Long submissionId) {
         User user = userRepository.findById(userId)
-            .orElseThrow(
-                () -> new CustomException("존재하지 않는 회원입니다.", ErrorCode.USER_NOT_FOUND));
+            .orElseThrow(() -> new CustomException("존재하지 않는 회원입니다.", ErrorCode.USER_NOT_FOUND));
 
-        Submission submission = submissionRepository.findById(submissionId)
-            .orElseThrow(
-                () -> new CustomException("프로그램 문제 정보가 잘못 되었습니다.", ErrorCode.BAD_REQUEST));
+        Submission submission = submissionRepository.findById(submissionId).orElseThrow(
+            () -> new CustomException("존재하지 않는 제출입니다.", ErrorCode.SUBMISSION_NOT_FOUND));
         // 해당 유저의 해당 프로그램 문제의 제출 모두 조회
         List<Submission> submissionHistories = submissionRepository.findAllByUserAndProgramProblemOrderByCreatedAtAsc(
-            user,
-            submission.getProgramProblem());
-        log.info("submissionHistories : {}", submissionHistories);
-        return new SubmissionListResponseDto(
-            submissionHistories.stream().map(history -> SubmissionResponseDto.from(history,
+            user, submission.getProgramProblem());
+
+        return new SubmissionListResponseDto(submissionHistories.stream().map(
+            history -> SubmissionResponseDto.from(history,
                 algorithmRepository.findAllAlgorithmsBySubmissionId(history.getId()))).toList());
     }
 
     @Override
+    @Transactional(readOnly = true)
     public UserSubmissionPageResponseDto getSubmissionMe(Long userId,
         UserSubmissionRequestDto userSubmissionRequestDto, Pageable pageable) {
         User user = userRepository.findById(userId)
-            .orElseThrow(
-                () -> new CustomException("존재하지 않는 회원입니다.", ErrorCode.USER_NOT_FOUND));
+            .orElseThrow(() -> new CustomException("존재하지 않는 회원입니다.", ErrorCode.USER_NOT_FOUND));
 
         Page<UserSubmissionResponseDto> submissionMeList = submissionRepository.findAllUserSubmissionList(
             user.getId(), userSubmissionRequestDto, pageable);
         return UserSubmissionPageResponseDto.from(submissionMeList);
     }
 
-//    @Override
-//    public TrendingSubmissionResponseDto getTrendingSubmission(String trendType) {
-//        return new TrendingSubmissionResponseDto(
-//            switch (trendType) {
-//                case "popular" -> submissionRepository.findMostParticipatedSubmissions();
-//                case "recent" -> submissionRepository.findRecentSubmissions();
-//                default -> submissionRepository.findHottestSubmissions();
-//            }
-//        );
-//    }
+    @Override
+    @Transactional(readOnly = true)
+    public TrendIdsResponseDto getTrendIds(String trendType) {
+        return switch (trendType) {
+            case "hot" ->
+                new TrendIdsResponseDto(submissionRepository.findHotSubmissionIds(), null);
+            case "recent" ->
+                new TrendIdsResponseDto(submissionRepository.findRecentSubmissionIds(), null);
+            case "join-in" ->
+                new TrendIdsResponseDto(null, submissionRepository.findTrendProgramProblemIds());
+            default ->
+                throw new CustomException("트렌드 조회 type이 잘못되었습니다.", ErrorCode.INVALID_PARAMETER);
+        };
+    }
 
 }
